Added delete functionality to helper tickets and removed query_entries.php

+++ assets/php/entries.php
@@ -1,6 +1,7 @@
 <script src="../../../js/jquery-3.5.1.js"></script>
 <?php include 'header.php';
       include_once ".env.php";
+	  include("functions/contactFunc.php");
 
 echo '<html>';
 	echo '<head>';
@@ -11,9 +12,7 @@ echo '<html>';
         crossorigin="anonymous">';
  	echo '</head>';
 	echo '<body>';
-		echo '<div class = "container product">';
-        	echo '<div class="alert alert-success" role="alert">';
-        	echo '</div>';
+		echo '<div class = "container product">';    	
         	echo '<div class="card">';
 				echo '<div class="card-header">';
 					echo '<div class="row">';
@@ -33,21 +32,55 @@ echo '<html>';
 				echo '</div>';
 			echo '</div>';
 			echo '<div id="content">';
-?>
-	<script>
-		function refresh_div(){
-			$.ajax({
-				type: 'get',
-				url: 'https://ec2-44-202-10-232.compute-1.amazonaws.com/assets/php/functions/query_entries.php',
-				success: function(data){
-					$('#content').html(data);
-				}
-			});
-		};
-		setInterval(function(){refresh_div();}, 500);
-	</script>
-<?php		echo '</div>';
-		echo '</div>';
-	echo '</body>';
+
+
+$dblink=db_iconnect("contact_us");
+$sql="Select * from `entries`";
+$result=$dblink->query($sql) or
+	die("Something went wrong with $sql<br>".$dblink->error);
+while($data=$result->fetch_array(MYSQLI_ASSOC)){
+echo '<div class="card">';
+	echo '<div class="card-body">';
+		echo '<form onSubmit="if(!confirm(\'Are you sure you would like to do this?\')){return false;}" method = "post" action = "">';
+			echo '<div class="row">';
+				echo '<div class="col-lg-3">';
+					echo '<label>Ticket: '.$data['ticket_id'].'</label>';
+					echo '<input type="hidden" name="ticket_id" value="'.$data['ticket_id'].'"/>';
+				echo '</div>';
+				echo '<div class="col-lg-3">';
+					echo '<label>Name: '.$data['first_name'].' '.$data['last_name'].'</label>';
+				echo '</div>';
+				echo '<div class="col-lg-3">';
+					echo '<label>Email: '.$data['email'].'</label>';
+				echo '</div>';
+				echo '<div class="col-lg-3">';
+					echo '<label>Phone Number: '.$data['phone'].'</label>';
+				echo '</div>';
+			echo '</div>';
+			echo '<div class="row">';
+				echo '<div class="col-lg-9">';
+					echo '<label>Comment: '.$data['comment'].'</label>';
+				echo '</div>';
+				echo '<div class="col-lg-3">';
+					echo '<input class="btn btn-danger" type="submit" role="button" name="delete" value="Delete"></input>';
+				echo '</div>';
+			echo '</div>';
+		echo '</form>';
+	echo '</div>';
+echo '</div>';
+}
+echo '</div>';
+echo '</div>';
+echo '</body>';
 echo '</html>';
+
+if(isset($_POST['delete'])){
+	$ticket_id = $_POST['ticket_id'];
+	$delete_sql = "DELETE from `entries` where ticket_id = '$ticket_id'";
+	$result2=$dblink->query($delete_sql)or
+		die("Something went wrong with $sql<br>".$dblink->error);
+	echo '<div class="alert alert-success" role="alert">'.$ticket_id.' has been deleted.</div>';
+}  
+
+header("Refresh:0, url=entries.php");
 ?>
